IGNITE-14861 code review changes
